- Pedestrian class & IPedestrian interface added - serving as base classes for all the upcoming actors - refactoring
